add decoode flag
