+++ tests/root_test.go
@@ -1,4 +1,4 @@
-package cmd
+package test
 
 import (
 	"fmt"
